requirements.txt
Django==5.2.5
gunicorn==23.0.0
whitenoise==6.8.2
protobuf==4.25.3
python-decouple==3.8
razorpay==1.4.2
twilio==9.7.1
Pillow==11.0.0
setuptools>=70.0.0
google-generativeai==0.7.2
